extract hooks to own files and directory
