+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContent.tsx
@@ -1,15 +1,13 @@
-import React, { ReactElement, useEffect, useRef, useState } from "react";
+import React, { ReactElement } from "react";
 import classnames from "classnames";
 import ReactDOM from "react-dom";
-import { usePopper } from "react-popper";
-import { useOnKeyDown } from "@jobber/hooks/useOnKeyDown";
-import { useRefocusOnActivator } from "@jobber/hooks/useRefocusOnActivator";
-import { useFocusTrap } from "@jobber/hooks/useFocusTrap";
 import styles from "./ComboboxContent.css";
 import { ComboboxSearch } from "./ComboboxSearch";
 import { ComboboxList } from "./ComboboxList";
 import { ComboboxContext } from "../../ComboboxProvider";
 import { ComboboxOption } from "../../Combobox.types";
+import { useComboboxContent } from "../../hooks/useComboboxContent";
+import { useComboboxAccessibility } from "../../hooks/useComboboxAccessibility";
 
 interface ComboboxContentProps {
   /**
@@ -121,147 +119,4 @@ export function ComboboxContent(props: ComboboxContentProps): JSX.Element {
     setSearchValue("");
     setOpen(false);
   }
-}
-
-function useComboboxAccessibility(
-  selectionCallback: (selection: ComboboxOption) => void,
-  filteredOptions: ComboboxOption[],
-  optionsListRef: React.RefObject<HTMLUListElement>,
-  open: boolean,
-  setOpen: (open: boolean) => void,
-  wrapperRef: React.RefObject<HTMLDivElement>,
-): {
-  popperRef: React.RefObject<HTMLDivElement>;
-  popperStyles: { [key: string]: React.CSSProperties };
-  attributes: { [key: string]: { [key: string]: string } | undefined };
-} {
-  const hasOptionsVisible = open && filteredOptions.length > 0;
-  const focusedIndex = useRef<number | null>(null);
-
-  useRefocusOnActivator(open);
-
-  const popperRef = useFocusTrap<HTMLDivElement>(open);
-  const { styles: popperStyles, attributes } = usePopper(
-    wrapperRef.current,
-    popperRef.current,
-    {
-      modifiers: [
-        {
-          name: "flip",
-          options: {
-            fallbackPlacements: ["top-start"],
-          },
-        },
-      ],
-      placement: "bottom-start",
-    },
-  );
-
-  useEffect(() => {
-    focusedIndex.current = null;
-  }, [open, filteredOptions.length]);
-
-  useEffect(() => {
-    if (open) {
-      popperRef.current?.addEventListener("keydown", handleContentKeydown);
-    }
-
-    return () => {
-      popperRef.current?.removeEventListener("keydown", handleContentKeydown);
-    };
-  }, [open, optionsListRef, filteredOptions]);
-
-  useOnKeyDown(() => {
-    if (open) {
-      setOpen(false);
-    }
-  }, "Escape");
-
-  function handleContentKeydown(event: KeyboardEvent) {
-    if (!hasOptionsVisible) return;
-
-    if (event.key === "Enter" || event.key === " ") {
-      const activeElementInList = optionsListRef.current?.contains(
-        document.activeElement,
-      );
-
-      if (!activeElementInList) return;
-
-      handleKeyboardSelection(event);
-    }
-    if (event.key === "ArrowDown") {
-      handleKeyboardNavigation(event, 1);
-    }
-    if (event.key === "ArrowUp") {
-      handleKeyboardNavigation(event, -1);
-    }
-  }
-
-  function handleKeyboardNavigation(event: KeyboardEvent, indexChange: number) {
-    const newIndex =
-      focusedIndex.current === null ? 0 : focusedIndex.current + indexChange;
-
-    if (newIndex < 0 || newIndex >= filteredOptions.length) return;
-
-    const optionElement = optionsListRef.current?.children[
-      newIndex
-    ] as HTMLElement;
-
-    optionElement?.focus();
-    // setFocusedIndex(newIndex);
-    focusedIndex.current = newIndex;
-    event.preventDefault();
-    event.stopPropagation();
-  }
-
-  function handleKeyboardSelection(event: KeyboardEvent) {
-    event.preventDefault();
-    event.stopPropagation();
-
-    if (focusedIndex.current !== null) {
-      selectionCallback(filteredOptions[focusedIndex.current]);
-    }
-  }
-
-  return {
-    popperRef,
-    popperStyles,
-    attributes,
-  };
-}
-
-function useComboboxContent(
-  options: ComboboxOption[],
-  open: boolean,
-): {
-  searchValue: string;
-  setSearchValue: React.Dispatch<React.SetStateAction<string>>;
-  setSelectedElement: React.Dispatch<React.SetStateAction<HTMLElement | null>>;
-  filteredOptions: ComboboxOption[];
-  optionsListRef: React.RefObject<HTMLUListElement>;
-} {
-  const [searchValue, setSearchValue] = useState<string>("");
-  const [selectedElement, setSelectedElement] = useState<HTMLElement | null>(
-    null,
-  );
-  const optionsListRef = useRef<HTMLUListElement>(null);
-  const filteredOptions = options.filter(option =>
-    option.label.toLowerCase().includes(searchValue.toLowerCase()),
-  );
-
-  useEffect(() => {
-    if (open && selectedElement) {
-      selectedElement?.scrollIntoView({
-        block: "nearest",
-      });
-    }
-  }, [open, selectedElement]);
-
-  return {
-    searchValue,
-    setSearchValue,
-    setSelectedElement,
-    filteredOptions,
-    optionsListRef,
-  };
 }
